requirements.txt
einops
hydra-core
jsonlines
lightning
numpy
pandas
timm
torch
tqdm
wandb
